Adding some basic support for switching betwen two screens.
