+++ .gitignore
@@ -1,2 +1,3 @@
 .expo
 node_modules
+.DS_Store
