+++ App.js
@@ -2,13 +2,39 @@ import { StatusBar } from 'expo-status-bar';
 import React , {useCallback} from 'react';
 import { Alert, Linking, StyleSheet, Text, View , Button, Image} from 'react-native';
 import SignInScreen from './Screens/SignInScreen'
+import { NavigationContainer } from '@react-navigation/native';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
+const Stack = createNativeStackNavigator();
 
-const url = "https://github.com/nwnatale24/JalapenoHotties"
-
-export default function App() {
+function DetailsScreen() {
   return (
-    <SignInScreen/>
+    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+      <Text>Details Screen</Text>
+    </View>
   );
 }
 
+function HomeScreen({ navigation }) {
+  return (
+    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+      <Text>Home Screen</Text>
+      <Button
+        title="Go to Details"
+        onPress={() => navigation.navigate('Details')}
+      />
+    </View>
+  );
+}
+
+export default function App() {
+  return (
+    <NavigationContainer>{
+      <Stack.Navigator>
+        <Stack.Screen name="Home" component={HomeScreen} />
+        <Stack.Screen name="Details" component={DetailsScreen} />
+      </Stack.Navigator>
+    
+    }</NavigationContainer>
+  );
+}
